components/google_calendar_service_api_client/pyproject.toml
[project]
name = "google-calendar-service-client"
version = "0.1.0"
description = "A client library for accessing Google Calendar Service"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.23.0,<0.29.0",
    "attrs>=22.2.0",
    "python-dateutil>=2.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["google_calendar_service_client"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
